versioon 3.2.2.11
PPA190; PPA191: Klassifikaatorite täiendused (täiendus) 
PPA037: Dokumendi muutmine uue dokumendi loomisega (arendus) 
PPA095: Töövoo automaatne peatumine kooskõlastamisel (arendus) 
PPA132: kasutaja nime juures puhkusel olemise kuvamine (arendus) 
PPA146: Lemmikute täiendused (arendus) 
PPA[nõudeta]: Asutuseülene kooskõlastamine nimetada asutustevaheliseks kooskõlastamiseks (täiendus) 
PPA179: AMRi liidese ja kasutaja andmete täiendused (arendus) 
DLSeadist: Klassifikaatorite lisamine ja kustutamine (1) (arendus) 
DLSeadist: Dokumendi liigi kustutamine (6) (arendus) 
DLSeadist: Uue dokumendi liigi lisamine (arendus) 
DLSeadist: Dokumendi liigi detailvaate salvestamine (6) (arendus) 
DLSeadist: Dokumendi liigi detailvaade - Metaandmed plokk - Objekti eemaldamine (4) (arendus) 
DLSeadist: Dokumendi liigi detailvaade - Metaandmed plokk - Andmevälja lisamine (4) (arendus) 
DLSeadist: Dokumendi liigi detailvaade - Metaandmed plokk - Andmeväljade grupi lisamine (4) (arendus) 
DLSeadist: Andmeväljade grupi detailvaade (arendus) 
DLSeadist: Andmeväljade grupi detailvaade - Andmevälja eemaldamine (3) (arendus) 
DLSeadist: Andmeväljade grupi detailvaade - Andmevälja lisamine (3) (arendus) 
DLSeadist: Dokumendi metaandmete ploki koostamine dokumendi liigi metaandmete alusel (5) (arendus) 
DLSeadist: Iseseisvate süsteemsete andmeväljade loomine (5) (arendus) 
DLSeadist: Süsteemne andmeväljade grupp "Registreerimisandmed" (5) (arendus) 
DLSeadist: Süsteemne andmeväljade grupp "Juurdepääsupiirang" (5) (arendus) 
DLSeadist: Süsteemne andmeväljade grupp "Dokumendi vastutaja" (5) (arendus) 
Koodiülevaatuseta integreeritud muudatused 
[nõudeta]: Dokumendi ekraanivormil töövoogude plokis eristada täitja ja kaastäitja tööülesanded (täiendus) 
EDHS-590 SIM DHS live täiendav soov: Töövood blokis peab olema mitmesammuline töövoog salvestamise/käivitamise järjekorras (täiendus) 
EDHS-591 SiM DHS live: saatmata dokumente näitab ühe dokumendi võrra rohkem kui on (funktsionaalsus vigane) 
Sama nimega kasutaja- ja kontaktgrupi loomine (pisivead) 
Töövoo peatamine ja tähtaeg (funktsionaalsus vigane) 
Dokumendi registreerimisnumbri genereerimine mustri alusel (arendus) 
EDHS-632 SIM DHS live: asjad ei ole ekraanivormil tähestikulises järjekorras ja delta teeb vaht suure-väikesel algustähel (funktsionaalsus 
vigane) 
EDHS-633 SIM DHS live: tööülesannet ei saa katkestada kuna ühel tööülesandel on varasem tähtaeg - vale teade ning vale kontroll 
(funktsionaalsus vigane) 
Klassifikaatori kustutamisel kontroll, et seda kusagil kasutusel ei ole (arendus) 
EDHS-647 SIM DHS live: Allkirjastamisel, pdf-i genereerimisel jääb faili nimesse alles ka doc-lühend 
Klassifikaatorite täiendustega arvestamine importimisel/eksportimisel (arendus) 
EDHS-652 SIM DHS live: ei saa lisada gruppi dokumendi õiguste haldusesse (funktsionaalsus vigane) 
EDHS-657 SIM DHS live: hiljem käivituval töövoole lubab delta panna varasema tähtaja (pisivead) 
Alusta uut dokumenti taust on paigast ära (optimeerimine) 
PPA030: Alusta uut dokumenti menüü täiendus ja DLSeadist: Põhimenüü punkti "Alusta uut dokumenti" ülesehituse ja reeglite muutmine (6) 
(pisivead) 
Süsteemi viga dokumendi vormilt lahkumisel (pisivead) 
DLSeadist: Eemaldada dokumendi liigi üldandmetest showValidityData (täiendus) 
mitte-süsteemne klassifikaator salvestatakse vigaselt reposse (funktsionaalsus kättesaamatu) 
PPA145: Vastutuse üle andmine
Skript olemasolevatel sarjadel dokumendi registreerimisnumbri mustri seadmiseks
Dokumendi viit ei tohi genereeruda tühjaks, lisada kontroll sarja detailvaatesse
DLSeadist: Klassifikaatoritega mitte seotud valikuväljade tähistamine + seotud reeglid



git-svn-id: https://svn.smit/tarkvara/edhs/trunk@107698 5fddf92a-9921-11dd-8d32-b999f448c456

+++ delta/source/java/ee/webmedia/alfresco/docdynamic/service/DocumentDynamicServiceImpl.java
@@ -1,5 +1,16 @@
 package ee.webmedia.alfresco.docdynamic.service;
 
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.ACCESS_RESTRICTION;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.CASE;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.FILE_CONTENTS;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.FILE_NAMES;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.FUNCTION;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.REG_DATE_TIME;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.REG_NUMBER;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.SERIES;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.SHORT_REG_NUMBER;
+import static ee.webmedia.alfresco.document.model.DocumentCommonModel.Props.VOLUME;
+
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -23,11 +34,16 @@ import ee.webmedia.alfresco.docadmin.service.DocumentAdminService;
 import ee.webmedia.alfresco.docadmin.service.DocumentType;
 import ee.webmedia.alfresco.docconfig.generator.SaveListener;
 import ee.webmedia.alfresco.docdynamic.model.DocumentDynamicModel;
+import ee.webmedia.alfresco.document.log.service.DocumentLogService;
 import ee.webmedia.alfresco.document.model.DocumentCommonModel;
+import ee.webmedia.alfresco.document.sendout.service.SendOutService;
 import ee.webmedia.alfresco.document.service.DocumentService;
+import ee.webmedia.alfresco.document.service.DocumentServiceImpl;
+import ee.webmedia.alfresco.document.service.EventsLoggingHelper;
 import ee.webmedia.alfresco.utils.MessageData;
 import ee.webmedia.alfresco.utils.MessageDataImpl;
 import ee.webmedia.alfresco.utils.MessageDataWrapper;
+import ee.webmedia.alfresco.utils.MessageUtil;
 import ee.webmedia.alfresco.utils.RepoUtil;
 import ee.webmedia.alfresco.utils.UnableToPerformMultiReasonException;
 
@@ -45,22 +61,48 @@ public class DocumentDynamicServiceImpl implements DocumentDynamicService, BeanF
     private GeneralService generalService;
     private DocumentService documentService;
     private DocumentAdminService documentAdminService;
+    private SendOutService sendOutService;
+    private DocumentLogService documentLogService;
 
     private BeanFactory beanFactory;
 
     @Override
     public NodeRef createDraft(String documentTypeId) {
+        QName type = DocumentDynamicModel.Types.DOCUMENT_DYNAMIC;
         DocumentType documentType = documentAdminService.getDocumentType(documentTypeId);
 
         Map<QName, Serializable> props = new HashMap<QName, Serializable>();
         props.put(DocumentDynamicModel.Props.DOCUMENT_TYPE_ID, documentType.getDocumentTypeId());
         props.put(DocumentDynamicModel.Props.DOCUMENT_TYPE_VERSION_NR, documentType.getLatestDocumentTypeVersion().getVersionNr());
 
+        // TODO temporary
+        Set<QName> aspects = generalService.getDefaultAspects(type);
+        for (QName docAspect : aspects) {
+            documentService.callbackAspectProperiesModifier(docAspect, props);
+        }
+
         NodeRef drafts = documentService.getDrafts();
-        NodeRef docRef = nodeService.createNode(drafts, DocumentCommonModel.Assocs.DOCUMENT, DocumentCommonModel.Assocs.DOCUMENT, DocumentDynamicModel.Types.DOCUMENT_DYNAMIC,
+        NodeRef docRef = nodeService.createNode(drafts, DocumentCommonModel.Assocs.DOCUMENT, DocumentCommonModel.Assocs.DOCUMENT, type,
                 props).getChildRef();
 
         return docRef;
+    }
+
+    @Override
+    public NodeRef copyDocument(DocumentDynamic document, Map<QName, Serializable> overriddenProperties, QName... ignoredProperty) {
+        NodeRef draftRef = createDraft(document.getDocumentTypeId());
+        Map<QName, Serializable> properties = RepoUtil.toQNameProperties(document.getNode().getProperties(), true);
+        // Override properties if needed
+        if (overriddenProperties != null) {
+            properties.putAll(overriddenProperties);
+        }
+        // Remove unnecessary properties
+        for (QName prop : ignoredProperty) {
+            properties.remove(prop);
+        }
+        nodeService.addProperties(draftRef, properties);
+
+        return draftRef;
     }
 
     @Override
@@ -89,6 +131,9 @@ public class DocumentDynamicServiceImpl implements DocumentDynamicService, BeanF
     @Override
     public void updateDocument(DocumentDynamic documentOriginal, List<String> saveListenerBeanNames) {
         DocumentDynamic document = documentOriginal.clone();
+        NodeRef docRef = document.getNodeRef();
+
+        document.setDraft(isDraft(docRef));
 
         ValidationHelperImpl validationHelper = new ValidationHelperImpl();
         for (String saveListenerBeanName : saveListenerBeanNames) {
@@ -103,13 +148,57 @@ public class DocumentDynamicServiceImpl implements DocumentDynamicService, BeanF
             saveListener.save(document);
         }
 
+        // If document is updated for the first time, add SEARCHABLE aspect to document and it's children files.
+        Map<String, Object> docProps = document.getNode().getProperties();
+        if (!nodeService.hasAspect(docRef, DocumentCommonModel.Aspects.SEARCHABLE)) {
+            nodeService.addAspect(docRef, DocumentCommonModel.Aspects.SEARCHABLE, null);
+            docProps.put(FILE_NAMES.toString(), documentService.getSearchableFileNames(docRef));
+            docProps.put(FILE_CONTENTS.toString(), documentService.getSearchableFileContents(docRef));
+            docProps.put(DocumentCommonModel.Props.SEARCHABLE_SEND_MODE.toString(), sendOutService.buildSearchableSendMode(docRef));
+        }
+
         LOG.info("updateDocument after validation and save listeners, before real saving: " + document);
-        generalService.setPropertiesIgnoringSystem(document.getNodeRef(), document.getNode().getProperties());
+        generalService.setPropertiesIgnoringSystem(docRef, docProps);
+        generalService.saveAddedAssocs(document.getNode());
+        { // update properties and log changes made in properties
+            DocumentServiceImpl.PropertyChangesMonitorHelper propertyChangesMonitorHelper = new DocumentServiceImpl.PropertyChangesMonitorHelper();// FIXME:
+            boolean propsChanged = propertyChangesMonitorHelper.setPropertiesIgnoringSystemAndReturnIfChanged(docRef, docProps //
+                    , FUNCTION, SERIES, VOLUME, CASE // location changes
+                    , REG_NUMBER, SHORT_REG_NUMBER, REG_DATE_TIME // registration changes
+                    , ACCESS_RESTRICTION // access restriction changed
+            );
+            if (!EventsLoggingHelper.isLoggingDisabled(document.getNode(), DocumentService.TransientProps.TEMP_LOGGING_DISABLED_DOCUMENT_METADATA_CHANGED)) {
+                if (document.isDraft()) {
+                    documentLogService.addDocumentLog(docRef, MessageUtil.getMessage("document_log_status_created"));
+                } else if (propsChanged) {
+                    documentLogService.addDocumentLog(docRef, MessageUtil.getMessage("document_log_status_changed"));
+                }
+            }
+        }
+
+        if (document.isDraft()) {
+            documentService.addPrivilegesBasedOnSeriesOnBackground(docRef);
+        }
     }
 
     @Override
-    public void deleteDocument(NodeRef docRef) {
+    public void deleteDocumentIfDraft(NodeRef docRef) {
+        LOG.info("deleteDocumentIfDraft document=" + docRef);
+        if (!nodeService.exists(docRef)) {
+            LOG.debug("Document does not exist, not deleting: " + docRef);
+            return;
+        }
+        if (!isDraft(docRef)) {
+            LOG.warn("Document is not a draft, not deleting: " + docRef);
+            return;
+        }
         nodeService.deleteNode(docRef);
+    }
+
+    @Override
+    public boolean isDraft(NodeRef docRef) {
+        NodeRef parentRef = nodeService.getPrimaryParent(docRef).getParentRef();
+        return DocumentCommonModel.Types.DRAFTS.equals(nodeService.getType(parentRef));
     }
 
     // START: setters
@@ -135,6 +224,14 @@ public class DocumentDynamicServiceImpl implements DocumentDynamicService, BeanF
 
     public void setDocumentAdminService(DocumentAdminService documentAdminService) {
         this.documentAdminService = documentAdminService;
+    }
+
+    public void setSendOutService(SendOutService sendOutService) {
+        this.sendOutService = sendOutService;
+    }
+
+    public void setDocumentLogService(DocumentLogService documentLogService) {
+        this.documentLogService = documentLogService;
     }
 
     @Override
